add headerBackTitleStyle screen option (#1083)
* add headerBackTitleStyle screen option

since the title is customizable, the back title should too.

* add documentation for `headerBackTitleStyle`

* make absence of headerBackTitle not affect return value

* merge styles so current scene can override

If there's no last scene, return the `headerBackTitleStyle` from the
current scene.
Otherwise return the `headerTitleStyle` from the last scene, and the
`headerBackTitleStyle` from the current scene, with the latter having
precedence.

* take headerBackTitleStyle from current scene

also the title style of the last scene is no longer taken into account